TT-225: Added controller tests for the certified-company variant
Modified the tests so they check that the certified-company pages will
correctly display the mobile phone installation message for each IDP
if the users have previously indicated on the select-phone page that they have a smart phone
but would prefer not to install apps.

When users describe whether they have a smart phone on the select-phone page, the select phone
variant controller will set a session variable called :reluctant_mob_installation.  The
certified-company form uses this value to determine whether it should render mobile app installation
warnings for each IDP (ie: May need installation, Will need installation or blank - which means
that the user doesn't need to install an app for the IDP).

For these tests, sessions are created which consider three states for the :reluctant_mob_installation
variable:
- true: meaning that the user would have indicated they have a smart phone but would prefer not to
install apps.
- false: meaning that the user has not indicated a 'reluctant yes', meaning they have either answered
Yes they have a smart phone or no they do not.
- not present: we do not assume that the previous controllers have even set this variable.

We have mocked an IDP to have a property :mobile_app_installation and checked that when
the user has said "yes, but I'd prefer not to" that the property value does appear and that when
they have specified any other answer that the property will not appear.

Authors: @kgarwood, @naseersharaf, cobainc0