TT-260: Fixing other documents page so it ignores stale answers from pressing back key
In the user journey, the following scenario causes an error:
1. Select passport and licence
2. Click Continue
3. Click Back
4. Now on the select phone page click the hyperlink "I don't have either of these documents"
5. Select has non-UK id
6. Click continue.
7. Click options for has mobile phone, has yes for install apps

The result will show the same IDPs as if a person had passport, a driver's licence and non-uk
ID document.  This is because the "Back" press causes the session to retain old selected
answers (ie: passport and licence).  Clicking the hyperlink means that the values on the
select_documents page are not refreshed and any answers for "true" for passport and licence
are retained until the IDP picker is chosen.

In other_identity_documents_controller, we have changed the behaviour of the select_other_documents
method so that no matter how the user arrives at the Other Documents page, it will replace rather
than combine 'documents' answers that may have been provided through the visit to the select_documents
page.  The result is that when the IDP page is reached, any true values for passport or licence will
not appear in form answers and will therefore not appear in the evidence that is used to filter IDPs.

We have done two tests so far:
(1) A test for selected_answer_store to ensure that successive sets to 'documents' will mean the latest
set of answers will overwrite the previous set of answers
(2) A feature test that imitates the scenario of having stale select_document answers which should be
ignored after visiting the other_identity_documents page.

Authors: @kgarwood, @cobainc0