TT-547: Generate nginx config for AB test
We only deploy branch builds when we intend to do AB testing on the
box. Some additional nginx configuration is required to do this so we
symlink it into the /etc/nginx/conf.d directory. This should be a
drop-in change so that removing the file when this package is
uninstalled will return nginx to its default behaviour.

Author: @anshulsirur