TT-563: Streamlined appearance of failed registration, will-not-work pages
The initial focus of the story was limited to making the will-not-work page look like the failed-registration page in the case where an RP is configured to continue on fail.  Some of this was agreeing on what kind of text should appear for scenarios where there would be a continue option and pages where it would not have one.  Some fragments of text originally appeared attached to other phrases that seemed to limit the potential re-use across pages that would either make use of them directly or would make use of them through a shared partial view (app/views/shared/_other_ways.html.rb).  As well as concerns about being able to re-use text, I had to also consider that the YML text message properties would sometimes contain HTML formatting codes.

Changes that were made to the pages and properties were changed in response to ongoing decisions about text would best fit the pages.

The spec pages were changed to correct minor formatting issues.

Authors: @kgarwood