TT-665: Reorganise ChooseACertifiedCompanyController tests
Controller logic and direct manipulation of the session is now tested
in the controller spec and the feature test is limited mainly to
testing user interaction and view rendering.

During the refactoring, it became apparent that the
ConfigurableJourneyController was adding an unnecessary
layer of abstraction for selecting whether the IDP interstitial
question or redirect warning page was displayed.

Author: @vixus0