TT671 - Fixed issue for disabling App Transparency AB test
- App transparency test was disabled by setting the control to 100%.
This was fine for new users but if previous users aleady had the app
transparency cookie set to variant they still saw the variant page.
Remedied by removing the cookie and ab logic from routes.rb
- Controller/Feature tests for the variant pages have been commented out
as they will fail to run when no ab test routes present.
- The variant version of controllers/templates have not been removed as
a decision on which version to keep has not been made yet

Author: @NasAshrafThoughtworks