BAU: Run linters by default in Rake
Currently, the linters are only run when explicitly called in the
`pre-commit.sh` script. We use Travis to provide us with feedback on
test status when reviewing pull requests but it only runs Rake. This
means we can miss linter errors when merging a PR which will proceed
to fail later in our build system.

Author: @vixus0