TT-721: Adding controller tests for redirect_to_idp_question page
Changing the controller logic so that when the form is valid it considers the yes or no response in deciding
on which route to go to next.  If the user answered 'yes' to the interstitial question, they go on to the
idp warning page as before.  But if they answer 'no', they go to a page we still have to design.  For now,
we've just stubbed the other route by using a convenient existing path like choose_a_certified_company_path.

Authors: @kgarwood, @javindo