TT-721: Adding new interstitial explanation property to IDP yml
When users answer no to the interstitial question they are forwarded to a page that
will contain IDP-specific information describing why the IDP won't work for them.  We
added a new property to the IDP YML files so the field values could be used for the new
page.

Authors: @kgarwood, @javindo