TT-809: Corrected some tests
In finishing this story and its related TT-721, we discovered
there was an error that was not being detected by our current test suites.
When we tried out a journey, the application generated errors complaining it
could not find the @idp variable whose fields would help populate the
idp_wont_work_for_you.html.erb file.

This error was not something either the pre-commit or acceptance tests were picking up,
so it indicated a gap in test coverage.  We added a new feature test to make sure
that when user answers no to the IDP interstitial question, the next page can be scraped
for the title of the idp-wont-work-for-you page. The test now helps identify if there
are problems rendering the page.

We also removed repetitive text that was appearing when users clicked "No" for the interstitial
question.

Authors: @kgarwood, @cobainc0,