TT-691: Log a warning when no selected IDP in session
Users can potentially reach certain pages (probably by directly hitting
the URL) that expect an IDP to be selected for the user's session.

A KeyError was being raised by Ruby which the user would
have seen as the something went wrong page. The KeyError was being
logged as a critical error in Sentry which was causing unnecessary
noise. Now, a warning is logged instead stating that no selected IDP
exists in the user's session. The user's experience is unchanged.

Authors: @vixus0 @javindo