TT-691 Fix nil transaction ID issue
Sentry was reporting quite a few errors where no transaction id was found for a user because their session had disappeared client side but it was attempting to look it up and erroring out when reporting to piwik. This change addresses this by catching and logging that error, then not proceeding to log to piwik.

Authors: @javindo @NasAshrafThoughtworks @cobainc0