TT-1119: Report the number of visible IDPs in piwik.
We'd like to find out how many uses need to scroll in order to view all
of the IDPs recommended to them.
In order to do this, we have added a new JS file, running only on the
'choose IDP' page, which counts the number of visible IDPs at page load
and reports it to piwik.

Authors: @michaelwalker @jakubmiarka