TT-988: Limit the number of services that appear.
The previous implementation of TT-988 allowed all enabled transactions to
appear in the transaction list. However this allowed in some test data that
should not be present.
The list now filters out transactions which are not explicitly in the
allowed in the transaction types display list in federation config.

Authors: @michaelwalker @jakubmiarka