BAU: Fix duplicate RPs appearing in the frontend.
Duplicate RPs were appearing in the front end due to accidently mutating
the list of name and homepage rps.
To fix this, we now use '+' instead of 'concat' to avoid mutating the
original object, and we clone the original list so that each correlator
has a unique version of the list.

Authors: @jakubmiarka, @hugh-emerson, @michaelwalker