TT-1257 Switch to strict matching of profiles.
Previously the profile matching logic was that the evidence provided had
to be a superset of at least one of the provided profiles.
This has been changed to require an exact match.
This allows us to not display IDPs based on performance, even if the user
has sufficient evidence for the IDP.

Authors: @hugh-emerson, @michaelwalker