TT-1347: improve data we send to Piwik.
Piwik can't handle simultaneous requests for a new visitor.
It mistakenly creates two separate visits.
This is happening now because of an A/B test on the start page sending
a second server-side request.
For now, we will stop sending requests from the start controller if the user
is in the loa1_shortened_journey A/B test. TT-1389 is to implement a better
long-term solution.

Authors: @hugh-emerson, @michaelWalker