TT-1134: Fix issue with LoaMatch functions for routes constraints
Issue is when a lambda is negated (!IsLoa1) then it no longer becomes a
lamba but a FalseClass. It's only by luck that it works as constraints
seem to treat a non lambda to evaluate as true and so add the necessary
routes. If two routes with the same name are added then it takes the
first one which is how it 'gets away with it'. Resolved by being more
explicit with the two lamba functions in LoaMatch.

Author: @NasAshrafThoughtworks