BAU: Implement trial function for AB testing
We'd like to test AB test variants in non-dev environments without enabling the test and issuing cookies to all visitors. This change will enable to set AB test to accept trial cookies and when the trial cookie is present and the value matches the experiment, it will enable user to see the variant without actually having the AB test cookie. This is immediately allowing the trial for the Clever Questions AB test.

Pair: @jakubmiarka @NasAshrafThoughtworks