TT-1606: Added piwik reporting of the AB test
Added start page reporting of AB test for the new clever questions test. This required changes to the start controllers due to piwik race condition and some changes to the tests.
When AB test is running from the start page, we stop reporting the event "The user has reached the start page" and report the AB test variant from the frontend (JS) to make sure the visits are tagged.

Author: @jakubmiarka