TT-1734: Simplify handling of IDP profiles
Switch from using individual profiles to recommend IDPs to using groups of
profiles called segments.

The intention is to allow the config to better match the terms used by the
analytics team, and to therefore improve the ease with which the config is
updated and maintained.
This change also opens up the option of automating the process of maintaining
this config in the future.

Authors: @jakubmiarka, @michaelwalker