TT-1755: Remove usages of 'forEach' from production code.
Array.Prototype.forEach is not supported in IE8, so we shouldn't be using
this in production code.
It has been replaced with jquery's 'each' method where jquery is already
used in the script, and with the native 'for' method otherwise.

I've also added semi-colon to some scripts where Intellij raises warnings.

Solo: @michaelwalker