TT-669: Spilt up ApplicationController into more managable sections.
ApplicationController was becoming very large and difficult to manage.

To help solve this, we created several 'partial controllers' and moved
methods out of ApplicationController into the new partial controllers.

Where the methods were only required in a few controllers, the new partial
controllers are now only included in those specific controllers.

Otherwise the partials are now required by ApplicationController.
In this case there is no conceptual difference in how the application runs
(ie all of the same methods are still available on application controller)
but the methods are packaged into meaningful modules, which should make the
application controller easier to understand.

Authors: @michaelwalker @callumknights