TT-1326: Replaced strings with references to YAML files
Our feature tests were using hard-coded strings for interactions and expectations. When any of the text changed in the YAML files, the tests failed and had to be updated. This is replacing the strings with references to the locales yaml. Also added Translation to the config, to make using translation simpler and consistent with the views (i.e. using t('path.to.text) instead of I18n.t(...)).

Solo: @jakubmiarka