HUB-2: IDP focused AB test
A new AB test is being introduced. The variant is a completely new page, which required a separate/new layout. It's using a small version of the IDP logos thus a new set of logos have been added and the IdentityProviderDisplayDecorator now includes these. The AB test is running on the start page for both LOAs, so to avoid Piwik race condition between start page, virtual start page and ab test reporting we're disabling the virtual start page reporting and adding the custom variables to the start page via JS client.

Pair: @jakubmiarka @callumknights