HUB-87: Custom failed registration page
Added a feature to show a custom failed registration page for a specific RP. This triggered by the
presence of a yml flag 'custom_fail_heading' in the RP yml file. This then renders a partial view
with the custom content, with the custom content itself specified in the RP yml file.

Also added a feature test that tests the custom failed registration page shows when it should do.