TT-1783: Visibility of Verify journey pages to search engines
In order to hide the /start page from Google Search results, we need to remove it manually from the Google Search Console. This commit adds a file to enable the Search Console for the Verify domain (www.signin.service.gov.uk). The Search Console will be tied to the special account created in the GDS G-Suite domain to manage it (details in the Jira ticket).