EID-636 Let the user choose between Verify and eIDAS on the Verify landing page
If you have eIDAS enabled then you will be asked to pick between verify or
eIDAS. If you do not you will be taken straight to the verify start page.
There will be a later story where services will be able to specify the service
they want with a hint.

Co-authored-by: Simon Worthington <simon.worthington@digital.cabinet-office.gov.uk>
Co-authored-by: James Howes <james.howes@digital.cabinet-office.gov.uk>