CVQ2-5 Call config service for RP translations
This commit adds functionality to call the Config Service whenever translations are required by the app. Calls to the Config Service are made on the `before_action` of any controller action, ensuring that translations are up-to-date for the RP related to the user's current journey. The frontend keeps an in-memory cache of these translations should the Config Service be unavailable at any point in the future, so that translations can still be served.

Solo: @karlbaker02