BAU: Fix bug in piwik reporting
Fix to remove bug in piwik reporting. The session variable user_followed_journey_hint was being set regardless of if journey hint was present on sign in journeys. This was happening after the initial virtual page was sent to piwik and was therefore only detectable on the outcome journey report.

solo: @rachelthecodesmith