HUB-113: Re-enable forcing a new piwik visit
Performance analyst team analysed the data from the day we forced
the new piwik visit for each RP request and they are happy to enable
as a default. We also identified a small issue when on start page we do
an additional JS reporting, the flag was still present, so added a fix
to remove it as soon as it gets used.

Solo: @jakubmiarka