HUB-190: Hide feedback form from integration.
The reason behind this is that since the IP whitelist has been removed the integration environment
has been visible to the public, including the feedback form, which on integration lacks any
zendesk config and merely throws an error when submitted. We decided to modify the frontend so that
when in integration if the feedback form is accessed a gov.uk styled error is shown to the user
stating that feedback is disabled in this environment.

To make these changes I have added a new view to the feedback folder, and made some changes to the
respective controller, that now checks the value of a new environment variable named
FEEDBACK_DISABLED, if true it renders the new view with the error message, if false renders the
feedback form as normal.

I have also added some unit tests for the additional functionality in the
controller, as well as a feature test to check the correct page is shown when feedback is disabled.