ZD-56753: Fix for incorrect view rendering
When a user is on variant, has a JS disabled and submits invalid form,
the select_documents_variant controller tries to re-render the page with
the error messages. However, we don't use that view and so it isn't present.
The fix is to render the control views (insteading of adding the variant view)

Solo: @jakubmiarka