HUB-195: Add controller and route to allow for redirect on start
In order to track users taking part in AB tests that begin in the Gov.UK
domain we need to be able to assign them a piwik user id cookie
before they are directed to the RP. This commit adds a redirect_to_rp
controller and route so that services specified in config will be
redirected to the rp before the start page. We will point at this route
from the variant of the Gov.uk A/B test. The RP will continue to point
at Verify allowing users to follow the normal route from then on. A
message will also be sent to Piwik from the new controller that a user
has begun an A/B test.

Co-Authored-by: Jakub Miarka <jakub.miarka@digital.cabinet-office.gov.uk>