ZD-64635: Constraint routes against bots
There are regular Sentry alerts when bots (?) are trying to query our URLs
with an appended ".json" at the end. This is erroring out as a MissingTemplate error and
surfacing in Sentry. This change is to constraint the "view" routes by html format, so
any non-html requests are treated as 404 without trying to render start.json etc...

solo: @jakubmiarka