CVQ2-177: remove the FederationTranslator
The FederationTranslator provided fallback to english functionality and errors when translations were missing. However, both of these features are already available in I18n by setting a fallback option to english and using the `::translate!' method to raise errors.

Removing this will make it easier to change some other classes to fix
the RP translation behaviour.