EID-885: Add ability to retry a failed eIDAS journey from the new error pages
The app will now reset an eIDAS journey in an error state due to an Authn or UAC failure. When the user clicks the link on the new error pages, frontend uses the new Policy endpoint to reset the session to SessionStartedState and then redirects the user to the Verify/eIDAS picker.

Author: @CharlesIC