CVQ2-177: use refreshing 'cache' in RpDisplayRepository
Implements a 'cache' inside the `RpDisplayRepository` that will refresh
the translations that support `RpDisplayData` every 30 minutes. The
refresh action on the cache allows for us to call off to config service
and fetch some new translations.

This removes the need for using the `before_action` in
`ApplicationController` to refresh a user's transaction's translations.

This also means that when we generate a transaction list each
transaction's display data will be refreshed if it has become stale.

One alternative to this approach might be to use the Rails.cache support
but I've decided to skip this for now as it looked a bit confusing with
the call into the translation service and it's not possible current to
cache the `display_data` object.