BAU: Adjust rubocop settings for Codacy
Rubocop config is only used by Codacy but not our pre-commit tool, so we never see any issues it comes up with. I've run rubocop locally and it came up with 3629 offenses, half of which (1826) are "line too long" violations. Clearly we don't stick to the default limit of 80 characters per line. Extending the limit to 140 reduces the line length violations to 208. This will make Codacy complain less.

Also disabled the "Missing magic comment `# frozen_string_literal: true`." check. This comment is meant to be added at the top of files when migrating from Ruby 2+ to Ruby 3+ so we don't need to worry about this right now.

Author: @CharlesIC