HUB-276: Pause page content change
Changed the old pause page with_user_session with updated content, to streamline
the user experience a bit and make the call to action a little clearer.
Essentially added a styled button that currently sends the user back to their RP,
this will be changed shortly but that is another story (HUB-264), as well as changing
the wording of the other content.