CVQ2-177: simplify the LoadingCache interface
The `LoadingCache` was being passed an object at creation that could
then be used to replace itself when a refesh event happenend. This was
a little awkward so I've changed the interface to loading cache so that
the way to refresh is passed to #fetch as a block. This is similar to
how `Hash#fetch` works.