HUB-278: Update journey hint cookie to support P&R
Changed the cookie structure to log RP as well and only keep the latest state.
Added a temporary piece of code to clean existing cookies in users browsers.
Kept the ATTEMPT and SUCCESS states in to stay backwards compatible and
also make sure we know if there ever was a success.