BAU: Bugfix for journey hints not being set
We've recently changed how we store selected IDP and selected Country in session.
At the same time we changed how we set the journey hint cookie using what's in the session.
When rebasing I missed the line where we read the value and used the old
syntax accessing the object. Which was returning nil and thus no hint was set.
This is to use the helper method to read from the session. It has a side
effect when processing authn response it is now strictly required to have a selected
IDP in the session. I think that's fine as it's part of the session anyway.
I updated the tests to setting the selected IDP.