HUB-368: Show user session timeout message based on not_on_or_after time
Added a new for for the timeout page, as well as relavent config for the new
content. Also added validation for the notOnOrAfter, which in frontend has
been renamed to assertion_expiry for clarity, so that the timeout page shows
when a user clicks 'submit' having exceeded notOnOrAfter.

Slight change to testing as well.