EID-1060: Show eIDAS scheme unavailable error page
Add a mechanism to handle the case when an eIDAS scheme may be unavailable to show the user the right error page.
Due to the nature of SAML we cannot handle the cases such as when the scheme returns a non-200 response to the browser. This is because the SAML authn request is POSTed bt the browser and then the response rendered, which is outside of our control. The only cases we can handle are those that originate within the Hub.

One such case is when the country's metadata is unavailable. This is usually caused by the country's metadata being down or the certs having changed. EidasSchemeUnavailableError is used to capture the types of error we want to classify and handle as "eIDAS scheme unavailable" case for an eIDAS journey. We can add more exception types to that list in the future if needed.

Also removed unused CEF Reference from the Stub API.

Author: @CharlesIC