EID-1269: Handle unexpected error response from Hub better
We had an incident recently where eIDAS was disabled in DR and the logs showed this message:
"Received 404 with error message: 'NONE', type: 'NONE' and id: 'NONE'"
This is not a great way to handle unexpected errors when sending a request to the Hub.

We already handle error responses from the Hub when they are known exceptions. These include anything to do with the logic and flow if a Verify journey. For any other issues, such as network errors or services being down, we throw an exception with a generic message. This message should reflect the fact that the problem is related to communication with the Hub rather than a problem with the user's journey.

Made expected error message in HubResponseHandler test to be regex so the new extended message is accepted.

Author: @CharlesIC