Add prometheus metrics endpoint
This commit adds a prometheus metrics endpoint to frontend.  It has
the same metrics that we expose to statsd.

I copied Metrics::ApiRequestReporter and
Metrics::ControllerActionReporter to Prometheus:: versions, but I
skipped Metrics::ResponseStatusReporter in favour of using the
Prometheus::Client::Collector middleware that instruments all http
requests.

It's expected that we don't expose the /metrics endpoint to the
outside world.