update gems
this updates gem dependencies.  One of the reasons I want to do this
is because when I tried to add the prometheus-client gem in #674 it
broke unless I did an update first.  However, I thought it was better
to split the update into its own PR and we can merge the prometheus
one later.

Two notable things:

 - the update caused linting to be fussier about indentation of the
   `private` keyword, so i had to update reporter_spec.rb
 - email_validator 2.0.0 is less strict (see
   balexand/email_validator#49) which caused feedback_form_spec.rb to
   fail; to avoid this I pinned us to 1.x