[prometheus] Use DirectFileStore when we run under puma
Prometheus metrics are collected and stored in-memory within a
process.  However when we run under puma there are multiple processes,
each of which collects their own metrics.  When we serve metrics on
the /metrics endpoint, we need to present all metrics for all
processes.

DirectFileStore is the solution to this: it stores metrics in a
directory which is shared between all worker processes.

This commit uses puma's `on_worker_boot` stanza to configure each
worker to use DirectFileStore with a shared temp dir.