use official prometheus client repo again
prometheus/client_ruby#95 has been merged (although not yet released
to rubygems) so we can use the official client git repo instead of the
gocardless fork.

This bumps us from 5dce3e5 (the latest commit on the gocardless
branch) to 460c2bb (the commit that merged the gocardless branch into
master) so it's a pretty minimal change.