BAU: Ban search robots from indexing pages
Setting the meta tag is not enough on error pages as Googlebot doesn't read it.
This uses the existing flag to append the header.

I also considered using the nginx rules, but it would be a bit more complex with the
conditions as we actually control the pages from frontend.