Remove unused JS
Following the migration to new infrastructure, this javascript is no
longer used.

Removing it will provide a small performance improvement and might make
it possible to tighten up the Content Security Policy (perhaps we can
remove `unsafe-inline` or `unsafe-eval` once this is gone).

We've already updated our privacy policy in #718