BAU: find out whether errors are recoverable
Add logging to test whether we have enough information to perform partial reecovery when users are shown the 'something has gone wrong' page. If this shows a good number of users could be helped, we'll redirect to a more helpful error page. Otherwise, this experiment should be reverted.

See https://trello.com/c/kl73OxUJ/241-improve-the-something-went-wrong-experience for more information.