BAU: test whether the service homepage is available
I think I've found a simpler way of getting the service homepage. It looks as if it ought to be present in the session. This would mean that we could just link to the homepage. This would avoid the need for translations, which seems to be the current primary source of failures. If this works, the only change required should be a fairly simple change to the 'something has gone wrong' template.