BAU: Try not to present users with a list of all services
When a user hits the 'something has gone wrong' page, we used to expect them to choose their service from a list of all the services. This change sees whether we know the service the user came from. If so, we give the user a simple link to click to go back to the start. Additionally, present the user with information on other ways to acieve their goal, in case Verify is repeatedly failing for them.