BAU: log error causes
Currently, we just log the raw exception. This makes it hard to work out what is causing the error pages to be shown. By indroducing a log that includes both the name of the partial and the error message, we should then be able to do a search in the logs that will show us what is causing the errors.