XFY6yswN: Change the Start again link on timeout page
We're inconsistent on when we send users back to the RP. There's a flag which
controls that (CONTINUE_ON_FAILED_REGISTRATION). However, it's used on the timeout page and the user
gets sent to the RP with error regardless.
This is to make sure users gets sent on non-success scenarios to RPs only if the flag
is enabled.