BAU: Fix checking for CSRF token
We've been seeing exception in the logs where the users get `TypeError: no implicit conversion of nil into String`.
This is caused by invalid session in combination with missing authenticity token (CSRF).
The current protect_from_forgery on the controller is not raising exception and therefore
keeps processing the request and failing further down. We have turned the logic around
(swapped `except` for `only`) so that the controller inherits the forgery protection
from the ApplicationController instead and we only overwrite the single method with skip.

Co-Authored-By: Phil Miller <phillip.miller@digital.cabinet-office.gov.uk>