EID-1732 When your service doesn't support eIDAS but the user thinks it should
This addresses two problems.

The first is that if a user replaces /start with /redirect-to-country on a service that doesn't support eIDAS the front end logs an error and Sentry starts showing red things.  This isn't really
an error in the sense of something that is going wrong because our code isn't handling a particular set of circumstances well so should only be logged as a warning.  In the course of investigating
this, @bjgill discovered that the test that tests this was passing even though it was failing.  The reason for this was that a specific session id was being set in the test but because the
session id didn't match what had been set-up, the something went wrong page was being displayed - this was the page that was expected so the test passed anyway.  Deleting the line that sets
the session id leaves the test to fail for the right reason (which was verified with the use of a temporary custom error page that could only be rendered from the place it was expected.

This incorporates the changes from EID-1733 which tests the controller directly whereas the feature test tests a user typing a URL in.