EID-1774 Implement eIDAS escape route (exploration)
In an ideal world we can use the referrer supplied with the request to give the user a route back to the sign-in picker they came from if their country isn't yet available as part of eidas.

Unfortunately, test-rp doesn't exactly mirror the way that the hub interacts with gov.uk so this adds some logging to enable us to see what the referrer looks like in real life.

Once the data's in, we can decide how to proceed.