v3vxhYgk: Enable frontend to expose certs for self-service
The self-service needs to query the config service to find out whether the new
certs are in use. In production it's straightforward as the app is already running
in the same VPC. However, it also needs to talk to integration, which is in a
different VPC/account.

This PR adds two new endpoints, which are featured-flagged and will be only enabled in
integration. One endpoint is for healthcheck and the other one for the actual certs.
It just passes the request on to the config service and returns the response.
There's an additional layer of authorization which checks for a key in the header which self-service is
going to send with all the requests - this is to avoid any unathorized flooding
of the config service.