HUH-74 Add last_successful_idp endpoint to HintController
We want to run a test where users are shown a direct link to log in with
the supplier they were last successful with. This is derived from their
cookie. Some JS is being written that will call this endpoint to find
out which IDP that was, and return some information which will be
useful.