HUH-74 Set selected IDP if successful idp hint set
We're testing a new flow where a journey hint is set on the interstitial
page if it is determined the user has previously signed in successfully.
The hint is the simple_id of the IDP, prefixed with `idp_`.

These changes allow that hint to be picked up, and the idp selected,
before redirecting the user to the idp.